fix lint errors

+++ src/treelearner/cuda/cuda_histogram_constructor.cu
@@ -303,11 +303,11 @@ __global__ void CUDAConstructDiscretizedHistogramDenseKernel(
       atomicAdd_system(feature_histogram_ptr + i, packed_grad_hess);
     }
   } else {
-    unsigned long long* feature_histogram_ptr = reinterpret_cast<unsigned long long*>(smaller_leaf_splits->hist_in_leaf) + partition_hist_start;
+    uint64_t* feature_histogram_ptr = reinterpret_cast<uint64_t*>(smaller_leaf_splits->hist_in_leaf) + partition_hist_start;
     for (unsigned int i = thread_idx; i < num_items_in_partition; i += num_threads_per_block) {
       const int32_t packed_grad_hess = shared_hist_packed[i];
       const int64_t packed_grad_hess_int64 = (static_cast<int64_t>(static_cast<int16_t>(packed_grad_hess >> 16)) << 32) | (static_cast<int64_t>(packed_grad_hess & 0x0000ffff));
-      atomicAdd_system(feature_histogram_ptr + i, (unsigned long long)(packed_grad_hess_int64));
+      atomicAdd_system(feature_histogram_ptr + i, (uint64_t)(packed_grad_hess_int64));
     }
   }
 }
@@ -368,11 +368,11 @@ __global__ void CUDAConstructDiscretizedHistogramSparseKernel(
       atomicAdd_system(feature_histogram_ptr + i, packed_grad_hess);
     }
   } else {
-    unsigned long long* feature_histogram_ptr = reinterpret_cast<unsigned long long*>(smaller_leaf_splits->hist_in_leaf) + partition_hist_start;
+    uint64_t* feature_histogram_ptr = reinterpret_cast<uint64_t*>(smaller_leaf_splits->hist_in_leaf) + partition_hist_start;
     for (unsigned int i = thread_idx; i < num_items_in_partition; i += num_threads_per_block) {
       const int32_t packed_grad_hess = shared_hist_packed[i];
       const int64_t packed_grad_hess_int64 = (static_cast<int64_t>(static_cast<int16_t>(packed_grad_hess >> 16)) << 32) | (static_cast<int64_t>(packed_grad_hess & 0x0000ffff));
-      atomicAdd_system(feature_histogram_ptr + i, (unsigned long long)(packed_grad_hess_int64));
+      atomicAdd_system(feature_histogram_ptr + i, (uint64_t)(packed_grad_hess_int64));
     }
   }
 }
